requirements.txt
discord.py==2.0.0
Pillow==9.5.0
